pyproject.toml
[project]
name = "typer-bot"
version = "0.1.0"
description = "Discord bot for football predictions"
requires-python = ">=3.10"
authors = [
    {name = "Typer Team"}
]
dependencies = [
    "discord-py>=2.3.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"